requirements.txt
pyyaml>=5.4.0
easydict
TensorBoardX
tensorboard
opencv-python
scipy
scikit-image
einops
h5py
nibabel
pandas
Pyarrow
matplotlib